project/requirements.txt
ansible
celery
django
djangorestframework
django-clear-cache
django-debug-toolbar
django-split-settings
django-webpack-loader
django_extensions
Fabric
#mysqlclient
psycopg2
python-memcached
redis
